Modified EPS subsystem module to conform entirely with pylint styling.
